refactor(desktop): consolidate file-icon system into renderer/lib/fileIcons
Move FileIcon / getFileIcon / resolveFileIconAssetUrl / loadFallthroughIcons
and the manifest typing into one shared module so the v2 sidebar trees, the
diff/pane chrome, and the chat/mention surfaces all draw from the same place
(v1 FilesView/utils now just re-exports it).

Also give unrecognized file types a real fallback icon in the Pierre trees:
loadFallthroughIcons now adds the Material default-file icon to the sprite and
remaps Pierre's generic `file` slot to it, matching what the non-tree FileIcon
surfaces already do via manifest.defaultIcon.
