+++ apps/desktop/src/renderer/components/Chat/ChatInterface/components/MentionPopover/MentionPopover.tsx
@@ -27,7 +27,7 @@ import {
 } from "react";
 import { HiMiniAtSymbol } from "react-icons/hi2";
 import { useDebouncedValue } from "renderer/hooks/useDebouncedValue";
-import { FileIcon } from "renderer/screens/main/components/WorkspaceView/RightSidebar/FilesView/utils";
+import { FileIcon } from "renderer/lib/fileIcons";
 
 const MAX_RESULTS = 20;
