+++ apps/desktop/src/renderer/lib/fileIcons/index.ts
@@ -1,0 +1,6 @@
+export { FileIcon } from "./FileIcon";
+export { getFileIcon } from "./getFileIcon";
+export { loadFallthroughIcons } from "./loadFallthroughIcons";
+export type { FileIconManifest } from "./manifest";
+export { fileIconManifest } from "./manifest";
+export { resolveFileIconAssetUrl } from "./resolveFileIconAssetUrl";
