+++ apps/desktop/src/renderer/lib/fileIcons/resolveFileIconAssetUrl.test.ts
