+++ apps/desktop/src/renderer/lib/fileIcons/resolveFileIconAssetUrl.ts
