--- apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/components/WorkspaceSidebar/utils/loadFallthroughIcons/index.ts
@@ -1,1 +1,0 @@
-export { loadFallthroughIcons } from "./loadFallthroughIcons";
