+++ apps/desktop/src/renderer/screens/main/components/WorkspaceView/RightSidebar/FilesView/utils/index.ts
@@ -1,3 +1,7 @@
-export { FileIcon } from "./FileIcon";
-export { getFileIcon } from "./file-icons";
-export { resolveFileIconAssetUrl } from "./resolveFileIconAssetUrl";
+// File-icon helpers now live in the shared `renderer/lib/fileIcons` module.
+// Re-exported here so existing v1 FilesView imports keep working.
+export {
+	FileIcon,
+	getFileIcon,
+	resolveFileIconAssetUrl,
+} from "renderer/lib/fileIcons";
